+++ hardpy/hardpy_panel/frontend/src/App.tsx
@@ -2,6 +2,7 @@
 // GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 import * as React from "react";
+import { useTranslation } from "react-i18next";
 
 import {
   Button,
@@ -17,7 +18,8 @@ import {
 } from "@blueprintjs/core";
 
 import StartStopButton from "./button/StartStop";
-import { SuiteList, TestRunI } from "./hardpy_test_view/SuiteList";
+import { TestRunI } from "./hardpy_test_view/SuiteList";
+import SuiteList from "./hardpy_test_view/SuiteList";
 import ProgressView from "./progress/ProgressView";
 import TestStatus from "./hardpy_test_view/TestStatus";
 import ReloadAlert from "./restart_alert/RestartAlert";
@@ -37,6 +39,7 @@ const WINDOW_WIDTH_THRESHOLDS = {
  * @returns {JSX.Element} The main application component.
  */
 function App(): JSX.Element {
+  const { t, i18n } = useTranslation();
   const [use_end_test_sound, setUseEndTestSound] = React.useState(false);
   const [use_debug_info, setUseDebugInfo] = React.useState(false);
 
@@ -92,7 +95,7 @@ function App(): JSX.Element {
     if (loading && rows.length === 0) {
       return (
         <Card style={{ marginTop: "60px" }}>
-          <H2>Establishing a connection... 🧐🔎</H2>
+          <H2>{t('app.connection')}</H2>
         </Card>
       );
     }
@@ -100,7 +103,7 @@ function App(): JSX.Element {
     if (state === "error") {
       return (
         <Card style={{ marginTop: "60px" }}>
-          <H2>Database connection error. 🙅🏽‍♀️🚫</H2>
+          <H2>{t('app.dbError')}</H2>
           {error && <p>{error.message}</p>}
         </Card>
       );
@@ -109,10 +112,11 @@ function App(): JSX.Element {
     if (rows.length === 0) {
       return (
         <Card style={{ marginTop: "60px" }}>
-          <H2>No entries in the database 🙅🏽‍♀️🚫</H2>
+          <H2>{t('app.noEntries')}</H2>
         </Card>
       );
     }
+
 
     /* Assume it is only one */
     const db_row = rows[0].doc as TestRunI;
@@ -174,18 +178,14 @@ function App(): JSX.Element {
       <Menu>
         <MenuItem
           shouldDismissPopover={false}
-          text={use_end_test_sound ? "Turn off the sound" : "Turn on the sound"}
+          text={use_end_test_sound ? t('app.soundOff') : t('app.soundOn')}
           icon={use_end_test_sound ? "volume-up" : "volume-off"}
           id="use_end_test_sound"
           onClick={() => setUseEndTestSound(!use_end_test_sound)}
         />
         <MenuItem
           shouldDismissPopover={false}
-          text={
-            use_debug_info
-              ? "Turn off the debug mode"
-              : "Turn on the debug mode"
-          }
+          text={use_debug_info ? t('app.debugOff') : t('app.debugOn')}
           icon={"bug"}
           id="use_debug_info"
           onClick={() => setUseDebugInfo(!use_debug_info)}
@@ -208,17 +208,17 @@ function App(): JSX.Element {
         <Navbar.Group align={Alignment.LEFT}>
           <Navbar.Heading id={"main-heading"}>
             <div className={"logo-smol"}></div>
-            {wide && (
-              <div>
-                <b>{ultrawide ? "HardPy Operator Panel" : "HardPy"}</b>
-              </div>
-            )}
+              {wide && (
+                <div>
+                  <b>{ultrawide ? t('app.title') : "HardPy"}</b>
+                </div>
+              )}
           </Navbar.Heading>
 
           {wide && <Navbar.Divider />}
 
           <Navbar.Heading id={"last-exec-heading"}>
-            <div>Last run:</div>
+            <div>{t('app.lastRun')}</div>
           </Navbar.Heading>
           <div id={"glob-test-status"}>
             <TestStatus status={lastRunStatus} />